rename `src/compile/util` to `src/compile/common`
