+++ src/compile/compile.ts
@@ -10,7 +10,7 @@ import {facetMixins} from './facet';
 import {compileLegends} from './legend';
 import {compileMark} from './mark/mark';
 import {compileScales} from './scale';
-import {applyConfig, FILL_STROKE_CONFIG} from './util';
+import {applyConfig, FILL_STROKE_CONFIG} from './common';
 import {extend} from '../util';
 
 import {LAYOUT} from '../data';
