+++ tsconfig.json
@@ -35,6 +35,7 @@
     "src/bin.ts",
     "src/channel.ts",
     "src/compile/axis.ts",
+    "src/compile/common.ts",
     "src/compile/compile.ts",
     "src/compile/config.ts",
     "src/compile/data.ts",
@@ -52,7 +53,6 @@
     "src/compile/scale.ts",
     "src/compile/stack.ts",
     "src/compile/time.ts",
-    "src/compile/util.ts",
     "src/data.ts",
     "src/encoding.ts",
     "src/enums.ts",
@@ -81,6 +81,7 @@
     "src/validate.ts",
     "src/vl.ts",
     "test/compile/axis.test.ts",
+    "test/compile/common.test.ts",
     "test/compile/compile.test.ts",
     "test/compile/data.test.ts",
     "test/compile/legend.test.ts",
@@ -93,7 +94,6 @@
     "test/compile/Model.test.ts",
     "test/compile/scale.test.ts",
     "test/compile/stack.test.ts",
-    "test/compile/util.test.ts",
     "test/fielddef.test.ts",
     "test/schema.test.ts",
     "test/type.test.ts",
